refactor(client): remove legacy xterm terminal renderer (closes #941)
Roadmap PR-5 (final): the labs @xterm/headless + custom React DOM renderer
is now the sole terminal, so the legacy xterm.js DOM renderer and its
satellite layer are removed after the default-next bake.

Deleted (26 files): components/Terminal.tsx; the IndexedDB cache layer
(terminal-state-cache, terminal-state-save-manager, terminal-chunk-writer,
terminal-restore); render-diagnostics; the xterm satellites (terminal-options,
terminal-mouse-protocol-guard, terminal-copy-on-select,
terminal-copy-prime-cleanup, terminal-key-handler); hooks/useTerminalWebSocket;
the renderer feature flag and its Settings toggle; the labs/terminal-compare
route; and all their tests.

Rewiring: SessionPage renders PocTerminalAdapter directly; main.tsx and
useSessionSideEffects drop the removed cache hooks; worker-websocket strips its
cache/diagnostics references and dead truncation-flag machinery. ConnectionStatus
and TerminalProps are rescued into labs/terminal-poc/terminal-contract.ts.

Dependencies: retired @xterm/xterm and the fit/serialize/web-links addons; kept
@xterm/headless. Client bundle drops ~347 kB raw (~89 kB gzip).

Note: worker-websocket's terminal transport is retained solely as MessagePanel's
Escape-to-PTY dependency, to be removed in the follow-up that rewires that path
through the PoC store.
