+++ packages/client/package.json
@@ -20,11 +20,7 @@
     "@radix-ui/react-dialog": "^1.1.15",
     "@tanstack/react-query": "^5.62.2",
     "@tanstack/react-router": "^1.86.1",
-    "@xterm/addon-fit": "^0.10.0",
-    "@xterm/addon-serialize": "^0.13.0",
-    "@xterm/addon-web-links": "^0.11.0",
     "@xterm/headless": "^5.5.0",
-    "@xterm/xterm": "^5.5.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "diff": "^8.0.2",
