+++ packages/client/src/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useAppWsState, useAppWsEvent } from './useAppWs';
-export { useTerminalWebSocket } from './useTerminalWebSocket';
