+++ packages/client/src/vite-env.d.ts
@@ -4,8 +4,3 @@ declare module '*.css' {
   const content: string;
   export default content;
 }
-
-declare module '@xterm/xterm/css/xterm.css' {
-  const content: string;
-  export default content;
-}
